refactor: Split Bird Families into bird_families_streamlit_html
Extract species-group coverage UI and prep helpers from rankings module;
add fragment tests and fix empty-taxonomy messaging when bundle exists.
