+++ explorer/app/streamlit/app_dashboard_shell.py
@@ -13,10 +13,8 @@ from explorer.app.streamlit.app_settings_ui import render_settings_tab
 from explorer.app.streamlit.checklist_stats_streamlit_html import run_checklist_stats_streamlit_fragment
 from explorer.app.streamlit.country_stats_streamlit_html import run_country_tab_streamlit_fragment
 from explorer.app.streamlit.maintenance_streamlit_html import run_maintenance_streamlit_tab_fragment
-from explorer.app.streamlit.rankings_streamlit_html import (
-    run_families_streamlit_tab_fragment,
-    run_rankings_streamlit_tab_fragment,
-)
+from explorer.app.streamlit.bird_families_streamlit_html import run_families_streamlit_tab_fragment
+from explorer.app.streamlit.rankings_streamlit_html import run_rankings_streamlit_tab_fragment
 from explorer.app.streamlit.streamlit_theme import inject_main_tab_panel_top_compact_css
 from explorer.app.streamlit.streamlit_ui_constants import NOTEBOOK_MAIN_TAB_LABELS
 from explorer.app.streamlit.yearly_summary_streamlit_html import run_yearly_summary_streamlit_fragment
